materialize-s3-iceberg: add alternative naming feature flag
Similar to the flag added to the materialize-iceberg connector, this
flag changes the table data paths in object storage to:
  <base_location>/<namespace>/<table>.<hash>

I brought the logic for table location naming into its own file for both
iceberg connectors to better be able to compare them.

There are some notable differences between them, unlike
materialize-iceberg: this connector allows hyphens in namespace and
table names and it always uses a nested layout with a directory for
the namespace.
